balkanized the single giant file and refactored batching
