+++ Source/AnchorGroupProvider.swift
@@ -1,0 +1,140 @@
+//
+//  AnchorGroupProvider.swift
+//  Anchorage
+//
+//  Created by Rob Visentin on 5/1/17.
+//
+//  Copyright 2016 Raizlabs and other contributors
+//  http://raizlabs.com/
+//
+//  Permission is hereby granted, free of charge, to any person obtaining
+//  a copy of this software and associated documentation files (the
+//  Software"), to deal in the Software without restriction, including
+//  without limitation the rights to use, copy, modify, merge, publish,
+//  distribute, sublicense, and/or sell copies of the Software, and to
+//  permit persons to whom the Software is furnished to do so, subject to
+//  the following conditions:
+//
+//  The above copyright notice and this permission notice shall be
+//  included in all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+//  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+//  MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+//  NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+//  LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+//  OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+//  WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+#if os(macOS)
+    import Cocoa
+#else
+    import UIKit
+#endif
+
+public protocol AnchorGroupProvider {
+
+    var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> { get }
+    var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> { get }
+    var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> { get }
+
+}
+
+extension AnchorGroupProvider {
+
+    public var edgeAnchors: EdgeAnchors {
+        return EdgeAnchors(horizontal: horizontalAnchors, vertical: verticalAnchors)
+    }
+
+}
+
+extension Alias.View: AnchorGroupProvider {
+
+    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> {
+        return AnchorPair(first: leadingAnchor, second: trailingAnchor)
+    }
+    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
+        return AnchorPair(first: topAnchor, second: bottomAnchor)
+    }
+    public var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> {
+        return AnchorPair(first: centerXAnchor, second: centerYAnchor)
+    }
+
+}
+
+extension Alias.ViewController: AnchorGroupProvider {
+
+    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> {
+        return AnchorPair(first: view.leadingAnchor, second: view.trailingAnchor)
+    }
+
+    #if os(macOS)
+    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
+        return AnchorPair(first: view.bottomAnchor, second: view.topAnchor)
+    }
+    #else
+    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
+    return AnchorPair(first: topLayoutGuide.bottomAnchor, second: bottomLayoutGuide.topAnchor)
+    }
+    #endif
+
+    public var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> {
+        return AnchorPair(first: view.centerXAnchor, second: view.centerYAnchor)
+    }
+
+}
+
+extension Alias.LayoutGuide: AnchorGroupProvider {
+
+    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> {
+        return AnchorPair(first: leadingAnchor, second: trailingAnchor)
+    }
+    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
+        return AnchorPair(first: topAnchor, second: bottomAnchor)
+    }
+    public var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> {
+        return AnchorPair(first: centerXAnchor, second: centerYAnchor)
+    }
+
+}
+
+// MARK: - EdgeAnchors
+
+public struct EdgeAnchors: LayoutAnchorType {
+
+    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor>
+    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor>
+    
+}
+
+// MARK: - Axis Group
+
+public struct AxisGroup {
+
+    public var first: NSLayoutConstraint
+    public var second: NSLayoutConstraint
+
+}
+
+// MARK: - EdgeGroup
+
+public struct EdgeGroup {
+
+    public var top: NSLayoutConstraint
+    public var leading: NSLayoutConstraint
+    public var bottom: NSLayoutConstraint
+    public var trailing: NSLayoutConstraint
+
+    public var horizontal: [NSLayoutConstraint] {
+        return [leading, trailing]
+    }
+
+    public var vertical: [NSLayoutConstraint] {
+        return [top, bottom]
+    }
+
+    public var all: [NSLayoutConstraint] {
+        return [top, leading, bottom, trailing]
+    }
+    
+}
