+++ Source/Anchorage.swift
@@ -28,58 +28,9 @@
 
 #if os(macOS)
     import Cocoa
-
-    public enum Alias {
-        public typealias View = NSView
-        public typealias ViewController = NSViewController
-        public typealias LayoutPriority = NSLayoutPriority
-        public typealias LayoutGuide = NSLayoutGuide
-
-        public static let LayoutPriorityRequired = NSLayoutPriorityRequired
-        public static let LayoutPriorityHigh = NSLayoutPriorityDefaultHigh
-        public static let LayoutPriorityLow = NSLayoutPriorityDefaultLow
-        public static let LayoutPriorityFittingSize = NSLayoutPriorityFittingSizeCompression
-    }
 #else
     import UIKit
-
-    public enum Alias {
-        public typealias View = UIView
-        public typealias ViewController = UIViewController
-        public typealias LayoutPriority = UILayoutPriority
-        public typealias LayoutGuide = UILayoutGuide
-
-        public static let LayoutPriorityRequired = UILayoutPriorityRequired
-        public static let LayoutPriorityHigh = UILayoutPriorityDefaultHigh
-        public static let LayoutPriorityLow = UILayoutPriorityDefaultLow
-        public static let LayoutPriorityFittingSize = UILayoutPriorityFittingSizeLevel
-    }
 #endif
-
-extension BinaryFloatingPoint {
-
-    static var exponentBias: Int {
-        return (1 << (Self.exponentBitCount - 1)) - 1
-    }
-
-    init<T: BinaryFloatingPoint>(_ value: T) {
-        assert(Self.radix == T.radix)
-        
-        let exp = value.exponent.toIntMax()
-        let convertedExp = UIntMax(exp + IntMax(Self.exponentBias))
-
-        let sig = value.significandBitPattern.toUIntMax()
-        let sigBitDiff = IntMax(Self.significandBitCount - T.significandBitCount)
-        let convertedSig = (sigBitDiff >= 0) ? (sig << UIntMax(sigBitDiff)) : (sig >> UIntMax(-sigBitDiff))
-
-        self.init(
-            sign: value.sign,
-            exponentBitPattern: RawExponent(convertedExp),
-            significandBitPattern: RawSignificand(convertedSig)
-        )
-    }
-    
-}
 
 public protocol LayoutAnchorType {}
 extension NSLayoutDimension : LayoutAnchorType {}
@@ -93,164 +44,164 @@ extension NSLayoutYAxisAnchor : LayoutAxisType {}
 // MARK: - Equality Constraints
 
 @discardableResult public func == <T: BinaryFloatingPoint>(lhs: NSLayoutDimension, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalToConstant: CGFloat(rhs)))
+    return finalize(constraint: lhs.constraint(equalToConstant: CGFloat(rhs)))
 }
 
 @discardableResult public func == (lhs: NSLayoutXAxisAnchor, rhs: NSLayoutXAxisAnchor) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalTo: rhs))
+    return finalize(constraint: lhs.constraint(equalTo: rhs))
 }
 
 @discardableResult public func == (lhs: NSLayoutYAxisAnchor, rhs: NSLayoutYAxisAnchor) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalTo: rhs))
+    return finalize(constraint: lhs.constraint(equalTo: rhs))
 }
 
 @discardableResult public func == (lhs: NSLayoutDimension, rhs: NSLayoutDimension) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalTo: rhs))
+    return finalize(constraint: lhs.constraint(equalTo: rhs))
 }
 
 @discardableResult public func == <T: NSLayoutDimension>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(equalTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func == <T: NSLayoutXAxisAnchor>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(equalTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func == <T: NSLayoutYAxisAnchor>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(equalTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(equalTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func == (lhs: NSLayoutDimension, rhs: LayoutExpression<NSLayoutDimension>) -> NSLayoutConstraint {
     if let anchor = rhs.anchor {
-        return activate(constraint: lhs.constraint(equalTo: anchor, multiplier: rhs.multiplier, constant: rhs.constant), withPriority: rhs.priority)
+        return finalize(constraint: lhs.constraint(equalTo: anchor, multiplier: rhs.multiplier, constant: rhs.constant), withPriority: rhs.priority)
     }
     else {
-        return activate(constraint: lhs.constraint(equalToConstant: rhs.constant), withPriority: rhs.priority)
+        return finalize(constraint: lhs.constraint(equalToConstant: rhs.constant), withPriority: rhs.priority)
     }
 }
 
 @discardableResult public func == (lhs: EdgeAnchors, rhs: EdgeAnchors) -> EdgeGroup {
-    return lhs.activate(constraintsEqualToEdges: rhs)
+    return lhs.finalize(constraintsEqualToEdges: rhs)
 }
 
 @discardableResult public func == (lhs: EdgeAnchors, rhs: LayoutExpression<EdgeAnchors>) -> EdgeGroup {
-    return lhs.activate(constraintsEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
+    return lhs.finalize(constraintsEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
 }
 
 @discardableResult public func == <T: LayoutAxisType, U: LayoutAxisType>(lhs: AnchorPair<T, U>, rhs: AnchorPair<T, U>) -> AxisGroup {
-    return lhs.activate(constraintsEqualToEdges: rhs)
+    return lhs.finalize(constraintsEqualToEdges: rhs)
 }
 
 @discardableResult public func == <T: LayoutAxisType, U: LayoutAxisType>(lhs: AnchorPair<T, U>, rhs: LayoutExpression<AnchorPair<T, U>>) -> AxisGroup {
-    return lhs.activate(constraintsEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
+    return lhs.finalize(constraintsEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
 }
 
 // MARK: - Inequality Constraints
 
 @discardableResult public func <= <T: BinaryFloatingPoint>(lhs: NSLayoutDimension, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualToConstant: CGFloat(rhs)))
+    return finalize(constraint: lhs.constraint(lessThanOrEqualToConstant: CGFloat(rhs)))
 }
 
 @discardableResult public func <= <T: NSLayoutDimension>(lhs: T, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualTo: rhs))
+    return finalize(constraint: lhs.constraint(lessThanOrEqualTo: rhs))
 }
 
 @discardableResult public func <= <T: NSLayoutXAxisAnchor>(lhs: T, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualTo: rhs))
+    return finalize(constraint: lhs.constraint(lessThanOrEqualTo: rhs))
 }
 
 @discardableResult public func <= <T: NSLayoutYAxisAnchor>(lhs: T, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualTo: rhs))
+    return finalize(constraint: lhs.constraint(lessThanOrEqualTo: rhs))
 }
 
 @discardableResult public func <= <T: NSLayoutDimension>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(lessThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func <= <T: NSLayoutXAxisAnchor>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(lessThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func <= <T: NSLayoutYAxisAnchor>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(lessThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(lessThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func <= (lhs: NSLayoutDimension, rhs: LayoutExpression<NSLayoutDimension>) -> NSLayoutConstraint {
     if let anchor = rhs.anchor {
-        return activate(constraint: lhs.constraint(lessThanOrEqualTo: anchor, multiplier: rhs.multiplier, constant: rhs.constant), withPriority: rhs.priority)
+        return finalize(constraint: lhs.constraint(lessThanOrEqualTo: anchor, multiplier: rhs.multiplier, constant: rhs.constant), withPriority: rhs.priority)
     }
     else {
-        return activate(constraint: lhs.constraint(lessThanOrEqualToConstant: rhs.constant), withPriority: rhs.priority)
+        return finalize(constraint: lhs.constraint(lessThanOrEqualToConstant: rhs.constant), withPriority: rhs.priority)
     }
 }
 
 @discardableResult public func <= (lhs: EdgeAnchors, rhs: EdgeAnchors) -> EdgeGroup {
-    return lhs.activate(constraintsLessThanOrEqualToEdges: rhs)
+    return lhs.finalize(constraintsLessThanOrEqualToEdges: rhs)
 }
 
 @discardableResult public func <= (lhs: EdgeAnchors, rhs: LayoutExpression<EdgeAnchors>) -> EdgeGroup {
-    return lhs.activate(constraintsLessThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
+    return lhs.finalize(constraintsLessThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
 }
 
 @discardableResult public func <= <T: LayoutAxisType, U: LayoutAxisType>(lhs: AnchorPair<T, U>, rhs: AnchorPair<T, U>) -> AxisGroup {
-    return lhs.activate(constraintsLessThanOrEqualToEdges: rhs)
+    return lhs.finalize(constraintsLessThanOrEqualToEdges: rhs)
 }
 
 @discardableResult public func <= <T: LayoutAxisType, U: LayoutAxisType>(lhs: AnchorPair<T, U>, rhs: LayoutExpression<AnchorPair<T, U>>) -> AxisGroup {
-    return lhs.activate(constraintsLessThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
+    return lhs.finalize(constraintsLessThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
 }
 
 @discardableResult public func >= <T: BinaryFloatingPoint>(lhs: NSLayoutDimension, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualToConstant: CGFloat(rhs)))
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualToConstant: CGFloat(rhs)))
 }
 
 @discardableResult public func >=<T: NSLayoutDimension>(lhs: T, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualTo: rhs))
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: rhs))
 }
 
 @discardableResult public func >=<T: NSLayoutXAxisAnchor>(lhs: T, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualTo: rhs))
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: rhs))
 }
 
 @discardableResult public func >=<T: NSLayoutYAxisAnchor>(lhs: T, rhs: T) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualTo: rhs))
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: rhs))
 }
 
 @discardableResult public func >= <T: NSLayoutDimension>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func >= <T: NSLayoutXAxisAnchor>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func >= <T: NSLayoutYAxisAnchor>(lhs: T, rhs: LayoutExpression<T>) -> NSLayoutConstraint {
-    return activate(constraint: lhs.constraint(greaterThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
+    return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: rhs.anchor!, constant: rhs.constant), withPriority: rhs.priority)
 }
 
 @discardableResult public func >= (lhs: NSLayoutDimension, rhs: LayoutExpression<NSLayoutDimension>) -> NSLayoutConstraint {
     if let anchor = rhs.anchor {
-        return activate(constraint: lhs.constraint(greaterThanOrEqualTo: anchor, multiplier: rhs.multiplier, constant: rhs.constant), withPriority: rhs.priority)
+        return finalize(constraint: lhs.constraint(greaterThanOrEqualTo: anchor, multiplier: rhs.multiplier, constant: rhs.constant), withPriority: rhs.priority)
     }
     else {
-        return activate(constraint: lhs.constraint(greaterThanOrEqualToConstant: rhs.constant), withPriority: rhs.priority)
+        return finalize(constraint: lhs.constraint(greaterThanOrEqualToConstant: rhs.constant), withPriority: rhs.priority)
     }
 }
 
 @discardableResult public func >= (lhs: EdgeAnchors, rhs: EdgeAnchors) -> EdgeGroup {
-    return lhs.activate(constraintsGreaterThanOrEqualToEdges: rhs)
+    return lhs.finalize(constraintsGreaterThanOrEqualToEdges: rhs)
 }
 
 @discardableResult public func >= (lhs: EdgeAnchors, rhs: LayoutExpression<EdgeAnchors>) -> EdgeGroup {
-    return lhs.activate(constraintsGreaterThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
+    return lhs.finalize(constraintsGreaterThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
 }
 
 @discardableResult public func >= <T: LayoutAxisType, U: LayoutAxisType>(lhs: AnchorPair<T, U>, rhs: AnchorPair<T, U>) -> AxisGroup {
-    return lhs.activate(constraintsGreaterThanOrEqualToEdges: rhs)
+    return lhs.finalize(constraintsGreaterThanOrEqualToEdges: rhs)
 }
 
 @discardableResult public func >= <T: LayoutAxisType, U: LayoutAxisType>(lhs: AnchorPair<T, U>, rhs: LayoutExpression<AnchorPair<T, U>>) -> AxisGroup {
-    return lhs.activate(constraintsGreaterThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
+    return lhs.finalize(constraintsGreaterThanOrEqualToEdges: rhs.anchor, constant: rhs.constant, priority: rhs.priority)
 }
 
 // MARK: - Priority
@@ -262,8 +213,6 @@ precedencegroup PriorityPrecedence {
 }
 
 infix operator ~: PriorityPrecedence
-
-// LayoutPriority
 
 @discardableResult public func ~ <T: BinaryFloatingPoint>(lhs: T, rhs: LayoutPriority) -> LayoutExpression<NSLayoutDimension> {
     return LayoutExpression(constant: CGFloat(lhs), priority: rhs)
@@ -278,8 +227,6 @@ infix operator ~: PriorityPrecedence
     expr.priority = rhs
     return expr
 }
-
-// MARK: Layout Expressions
 
 @discardableResult public func * <T: BinaryFloatingPoint>(lhs: NSLayoutDimension, rhs: T) -> LayoutExpression<NSLayoutDimension> {
     return LayoutExpression(anchor: lhs, multiplier: CGFloat(rhs))
@@ -351,362 +298,28 @@ infix operator ~: PriorityPrecedence
     return expr
 }
 
-// MARK: - LayoutExpression
-
-public struct LayoutExpression<T: LayoutAnchorType> {
-
-    public var anchor: T?
-    public var constant: CGFloat
-    public var multiplier: CGFloat
-    public var priority: LayoutPriority
-
-    fileprivate init(anchor: T? = nil, constant: CGFloat = 0.0, multiplier: CGFloat = 1.0, priority: LayoutPriority = .required) {
-        self.anchor = anchor
-        self.constant = constant
-        self.multiplier = multiplier
-        self.priority = priority
-    }
-
-}
-
-// MARK: - LayoutPriority
-
-public enum LayoutPriority: ExpressibleByFloatLiteral, ExpressibleByIntegerLiteral, Equatable {
-
-    case required
-    case high
-    case low
-    case fittingSize
-    case custom(Alias.LayoutPriority)
-
-    public var value: Alias.LayoutPriority {
-        switch self {
-        case .required: return Alias.LayoutPriorityRequired
-        case .high: return Alias.LayoutPriorityHigh
-        case .low: return Alias.LayoutPriorityLow
-        case .fittingSize: return Alias.LayoutPriorityFittingSize
-        case .custom(let priority): return priority
-        }
-    }
-
-    public init(floatLiteral value: Alias.LayoutPriority) {
-        self.init(value)
-    }
-
-    public init(integerLiteral value: Int) {
-        self.init(value)
-    }
-
-    public init(_ value: Int) {
-        self = .custom(Alias.LayoutPriority(value))
-    }
-
-    public init<T: BinaryFloatingPoint>(_ value: T) {
-        self = .custom(Alias.LayoutPriority(value))
-    }
-
-}
-
-public func == (lhs: LayoutPriority, rhs: LayoutPriority) -> Bool {
-    return lhs.value == rhs.value
-}
-
-public func + <T: BinaryFloatingPoint>(lhs: LayoutPriority, rhs: T) -> LayoutPriority {
-    return .custom(lhs.value + Alias.LayoutPriority(CGFloat(rhs)))
-}
-
-public func + <T: BinaryFloatingPoint>(lhs: T, rhs: LayoutPriority) -> LayoutPriority {
-    return .custom(Alias.LayoutPriority(CGFloat(lhs)) + rhs.value)
-}
-
-public func - <T: BinaryFloatingPoint>(lhs: LayoutPriority, rhs: T) -> LayoutPriority {
-    return .custom(lhs.value - Alias.LayoutPriority(CGFloat(rhs)))
-}
-
-public func - <T: BinaryFloatingPoint>(lhs: T, rhs: LayoutPriority) -> LayoutPriority {
-    return .custom(Alias.LayoutPriority(CGFloat(lhs)) - rhs.value)
-}
-
-// MARK: - EdgeAnchorsProvider
-
-public protocol AnchorGroupProvider {
-
-    var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> { get }
-    var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> { get }
-    var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> { get }
-
-}
-
-extension AnchorGroupProvider {
-
-    public var edgeAnchors: EdgeAnchors {
-        return EdgeAnchors(horizontal: horizontalAnchors, vertical: verticalAnchors)
-    }
-
-}
-
-extension Alias.View: AnchorGroupProvider {
-
-    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> {
-        return AnchorPair(first: leadingAnchor, second: trailingAnchor)
-    }
-    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: topAnchor, second: bottomAnchor)
-    }
-    public var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: centerXAnchor, second: centerYAnchor)
-    }
-
-}
-
-extension Alias.ViewController: AnchorGroupProvider {
-
-    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> {
-        return AnchorPair(first: view.leadingAnchor, second: view.trailingAnchor)
-    }
-
-#if os(macOS)
-    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: view.bottomAnchor, second: view.topAnchor)
-    }
-#else
-    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: topLayoutGuide.bottomAnchor, second: bottomLayoutGuide.topAnchor)
-    }
-#endif
-
-    public var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: view.centerXAnchor, second: view.centerYAnchor)
-    }
-
-}
-
-extension Alias.LayoutGuide: AnchorGroupProvider {
-
-    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor> {
-        return AnchorPair(first: leadingAnchor, second: trailingAnchor)
-    }
-    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: topAnchor, second: bottomAnchor)
-    }
-    public var centerAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutYAxisAnchor> {
-        return AnchorPair(first: centerXAnchor, second: centerYAnchor)
-    }
-
-}
-
-// MARK: - AnchorPair
-
-public struct AnchorPair<T: LayoutAxisType, U: LayoutAxisType>: LayoutAnchorType {
-
-    public var first: T
-    public var second: U
-
-    fileprivate init(first: T, second: U) {
-        self.first = first
-        self.second = second
-    }
-
-    fileprivate func activate(constraintsEqualToEdges anchor: AnchorPair<T, U>?, constant c: CGFloat = 0.0, priority: LayoutPriority = .required) -> AxisGroup {
-        let builder = ConstraintBuilder(horizontal: ==, vertical: ==)
-        return constraints(forAnchors: anchor, constant: c, priority: priority, builder: builder)
-    }
-
-    fileprivate func activate(constraintsLessThanOrEqualToEdges anchor: AnchorPair<T, U>?, constant c: CGFloat = 0.0, priority: LayoutPriority = .required) -> AxisGroup {
-        let builder = ConstraintBuilder(leading: <=, top: <=, trailing: >=, bottom: >=, centerX: <=, centerY: <=)
-        return constraints(forAnchors: anchor, constant: c, priority: priority, builder: builder)
-    }
-
-    fileprivate func activate(constraintsGreaterThanOrEqualToEdges anchor: AnchorPair<T, U>?, constant c: CGFloat = 0.0, priority: LayoutPriority = .required) -> AxisGroup {
-        let builder = ConstraintBuilder(leading: >=, top: >=, trailing: <=, bottom: <=, centerX: >=, centerY: >=)
-        return constraints(forAnchors: anchor, constant: c, priority: priority, builder: builder)
-    }
-
-    fileprivate func constraints(forAnchors anchors: AnchorPair<T, U>?, constant c: CGFloat, priority: LayoutPriority, builder: ConstraintBuilder) -> AxisGroup {
-        guard let anchors = anchors else {
-            preconditionFailure("Encountered nil edge anchors, indicating internal inconsistency of this API.")
-        }
-
-        switch (first, anchors.first, second, anchors.second) {
-        // Leading, trailing
-        case let (firstX as NSLayoutXAxisAnchor, otherFirstX as NSLayoutXAxisAnchor,
-                  secondX as NSLayoutXAxisAnchor, otherSecondX as NSLayoutXAxisAnchor):
-            return AxisGroup(first: builder.leadingBuilder(firstX, otherFirstX + c ~ priority),
-                             second: builder.trailingBuilder(secondX, otherSecondX - c ~ priority))
-        //Top, bottom
-        case let (firstY as NSLayoutYAxisAnchor, otherFirstY as NSLayoutYAxisAnchor,
-                  secondY as NSLayoutYAxisAnchor, otherSecondY as NSLayoutYAxisAnchor):
-            return AxisGroup(first: builder.topBuilder(firstY, otherFirstY + c ~ priority),
-                             second: builder.bottomBuilder(secondY, otherSecondY - c ~ priority))
-        //CenterX, centerY
-        case let (firstX as NSLayoutXAxisAnchor, otherFirstX as NSLayoutXAxisAnchor,
-                  firstY as NSLayoutYAxisAnchor, otherFirstY as NSLayoutYAxisAnchor):
-            return AxisGroup(first: builder.centerXBuilder(firstX, otherFirstX + c ~ priority),
-                             second: builder.centerYBuilder(firstY, otherFirstY + c ~ priority))
-        default:
-            preconditionFailure("Layout axes of constrained anchors must match.")
-        }
-    }
-}
-
-// MARK: - EdgeAnchors
-
-public struct EdgeAnchors: LayoutAnchorType {
-
-    public var horizontalAnchors: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor>
-    public var verticalAnchors: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor>
-
-    fileprivate init(horizontal: AnchorPair<NSLayoutXAxisAnchor, NSLayoutXAxisAnchor>, vertical: AnchorPair<NSLayoutYAxisAnchor, NSLayoutYAxisAnchor>) {
-        self.horizontalAnchors = horizontal
-        self.verticalAnchors = vertical
-    }
-
-    fileprivate func activate(constraintsEqualToEdges anchor: EdgeAnchors?, constant c: CGFloat = 0.0, priority: LayoutPriority = .required) -> EdgeGroup {
-        let builder = ConstraintBuilder(horizontal: ==, vertical: ==)
-        return constraints(forAnchors: anchor, constant: c, priority: priority, builder: builder)
-    }
-
-    fileprivate func activate(constraintsLessThanOrEqualToEdges anchor: EdgeAnchors?, constant c: CGFloat = 0.0, priority: LayoutPriority = .required) -> EdgeGroup {
-        let builder = ConstraintBuilder(leading: <=, top: <=, trailing: >=, bottom: >=, centerX: <=, centerY: <=)
-        return constraints(forAnchors: anchor, constant: c, priority: priority, builder: builder)
-    }
-
-    fileprivate func activate(constraintsGreaterThanOrEqualToEdges anchor: EdgeAnchors?, constant c: CGFloat = 0.0, priority: LayoutPriority = .required) -> EdgeGroup {
-        let builder = ConstraintBuilder(leading: >=, top: >=, trailing: <=, bottom: <=, centerX: >=, centerY: >=)
-        return constraints(forAnchors: anchor, constant: c, priority: priority, builder: builder)
-    }
-
-    fileprivate func constraints(forAnchors anchors: EdgeAnchors?, constant c: CGFloat, priority: LayoutPriority, builder: ConstraintBuilder) -> EdgeGroup {
-        guard let anchors = anchors else {
-            preconditionFailure("Encountered nil edge anchors, indicating internal inconsistency of this API.")
-        }
-
-        let horizontalConstraints = horizontalAnchors.constraints(forAnchors: anchors.horizontalAnchors, constant: c, priority: priority, builder: builder)
-        let verticalConstraints = verticalAnchors.constraints(forAnchors: anchors.verticalAnchors, constant: c, priority: priority, builder: builder)
-        return EdgeGroup(top: verticalConstraints.first,
-                         leading: horizontalConstraints.first,
-                         bottom: verticalConstraints.second,
-                         trailing: horizontalConstraints.second)
-    }
-
-}
-
-// MARK: - EdgeGroup
-
-public struct EdgeGroup {
-
-    public var top: NSLayoutConstraint
-    public var leading: NSLayoutConstraint
-    public var bottom: NSLayoutConstraint
-    public var trailing: NSLayoutConstraint
-
-    public var horizontal: [NSLayoutConstraint] {
-        return [leading, trailing]
-    }
-
-    public var vertical: [NSLayoutConstraint] {
-        return [top, bottom]
-    }
-
-    public var all: [NSLayoutConstraint] {
-        return [top, leading, bottom, trailing]
-    }
-
-}
-
-// MARK: - Axis Group
-
-public struct AxisGroup {
-
-    public var first: NSLayoutConstraint
-    public var second: NSLayoutConstraint
-
-}
-
-// MARK: - ConstraintBuilder
-
-private struct ConstraintBuilder {
-
-    typealias Horizontal = (NSLayoutXAxisAnchor, LayoutExpression<NSLayoutXAxisAnchor>) -> NSLayoutConstraint
-    typealias Vertical = (NSLayoutYAxisAnchor, LayoutExpression<NSLayoutYAxisAnchor>) -> NSLayoutConstraint
-
-    var topBuilder: Vertical
-    var leadingBuilder: Horizontal
-    var bottomBuilder: Vertical
-    var trailingBuilder: Horizontal
-    var centerYBuilder: Vertical
-    var centerXBuilder: Horizontal
-
-    init(horizontal: @escaping Horizontal, vertical: @escaping Vertical) {
-        topBuilder = vertical
-        leadingBuilder = horizontal
-        bottomBuilder = vertical
-        trailingBuilder = horizontal
-        centerYBuilder = vertical
-        centerXBuilder = horizontal
-    }
-
-    init(leading: @escaping Horizontal, top: @escaping Vertical, trailing: @escaping Horizontal,
-         bottom: @escaping Vertical, centerX: @escaping Horizontal, centerY: @escaping Vertical) {
-        leadingBuilder = leading
-        topBuilder = top
-        trailingBuilder = trailing
-        bottomBuilder = bottom
-        centerYBuilder = centerY
-        centerXBuilder = centerX
-    }
-
-}
-
-// MARK: - Constraint Activation
-
-private func activate(constraint theConstraint: NSLayoutConstraint, withPriority priority: LayoutPriority = .required) -> NSLayoutConstraint {
-    // Only disable autoresizing constraints on the LHS item, which is the one definitely intended to be governed by Auto Layout
-    if let first = theConstraint.firstItem as? Alias.View {
-        first.translatesAutoresizingMaskIntoConstraints = false
-    }
-
-    theConstraint.priority = priority.value
-    if BatchingConstraintBuilder.buildInactiveConstraints {
-        BatchingConstraintBuilder.pendingConstraints.append(theConstraint)
-    }
-    else {
-        theConstraint.isActive = true
-    }
-
-    return theConstraint
-}
-
-// MARK: - Batched Constraint Creation
-
-private enum BatchingConstraintBuilder {
-
-    static var buildInactiveConstraints = false
-    static var pendingConstraints: [NSLayoutConstraint] = []
-
-    /// Returns the pending constraints.
-    /// - Postcondition: After calling this method, the value of `pendingConstraints` will be the empty array.
-    /// - Returns: the pending constraints.
-    static func flushPendingConstraints() -> [NSLayoutConstraint] {
-        defer { pendingConstraints = [] }
-        return pendingConstraints
-    }
-
-}
+// MARK: - Batching
 
 /// Any Anchorage constraints created inside the passed closure are returned in the array.
 ///
+/// - Precondition: Can't be called inside or simultaneously with another batch. Batches cannot be nested.
 /// - Parameter active: Whether the created constraints should be active when they are returned.
 /// - Parameter closure: A closure that runs some Anchorage expressions.
 /// - Returns: An array of new `NSLayoutConstraint`s.
-public func batch(active: Bool, closure: () -> Void) -> [NSLayoutConstraint] {
-    precondition(BatchingConstraintBuilder.pendingConstraints.isEmpty)
-    BatchingConstraintBuilder.buildInactiveConstraints = true
-    closure()
-    BatchingConstraintBuilder.buildInactiveConstraints = false
-    let constraints = BatchingConstraintBuilder.flushPendingConstraints()
-    if active {
-        NSLayoutConstraint.activate(constraints)
+@discardableResult public func batch(active: Bool, closure: () -> Void) -> [NSLayoutConstraint] {
+    precondition(currentBatch == nil)
+    defer {
+        currentBatch = nil
     }
-    return constraints
+
+    let batch = ConstraintBatch()
+    currentBatch = batch
+
+    closure()
+
+    if active {
+        batch.activate()
+    }
+
+    return batch.constraints
 }
